app/requirements.txt
clip==1.0
langchain==0.0.352
openai==1.6.1
pinecone-client==2.2.4
streamlit==1.29.0
torch==2.1.2
